+++ packages/lumx-react/src/components/uploader/Uploader.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { commonTestsSuiteRTL } from '@lumx/react/testing/utils';
 
-import { render } from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
 import { getByClassName, queryByClassName } from '@lumx/react/testing/utils/queries';
 import userEvent from '@testing-library/user-event';
 import { mdiPlus } from '@lumx/icons';
@@ -29,9 +29,11 @@ const setup = (propsOverride: SetupProps = {}) => {
 describe(`<${Uploader.displayName}>`, () => {
     describe('Props', () => {
         it('should render default', () => {
-            const { uploader } = setup();
+            const label = 'Label';
+            const { uploader } = setup({ label });
 
             expect(uploader).toHaveClass(CLASSNAME);
+            expect(uploader).toBe(screen.queryByRole('button', { name: label }));
             expect(uploader).toHaveClass('lumx-uploader--aspect-ratio-horizontal');
             expect(uploader).toHaveClass('lumx-uploader--size-xl');
             expect(uploader).toHaveClass('lumx-uploader--theme-light');
@@ -41,12 +43,6 @@ describe(`<${Uploader.displayName}>`, () => {
         it('should render icon', () => {
             const { icon } = setup({ icon: mdiPlus });
             expect(icon).toBeInTheDocument();
-        });
-
-        it('should render label', () => {
-            const { label } = setup({ label: 'Label' });
-            expect(label).toBeInTheDocument();
-            expect(label).toHaveTextContent('Label');
         });
 
         it('should render variant circle', () => {
@@ -75,17 +71,29 @@ describe(`<${Uploader.displayName}>`, () => {
     });
 
     describe('Events', () => {
-        const onClick = jest.fn();
-
-        beforeEach(() => {
-            onClick.mockClear();
-        });
-
         it('should trigger `onClick` when clicked', async () => {
+            const onClick = jest.fn();
             const { uploader } = setup({ onClick });
 
             await userEvent.click(uploader);
+            expect(onClick).toHaveBeenCalled();
+        });
 
+        it('should trigger `onClick` when pressing Enter or Escape', async () => {
+            const onClick = jest.fn();
+            const { uploader } = setup({ onClick });
+
+            await userEvent.tab();
+            expect(uploader).toHaveFocus();
+
+            // Activate with Enter
+            await userEvent.keyboard('[Enter]');
+            expect(onClick).toHaveBeenCalled();
+
+            onClick.mockClear();
+
+            // Activate with Space
+            await userEvent.keyboard('[Space]');
             expect(onClick).toHaveBeenCalled();
         });
     });
@@ -95,5 +103,6 @@ describe(`<${Uploader.displayName}>`, () => {
         baseClassName: CLASSNAME,
         forwardClassName: 'uploader',
         forwardAttributes: 'uploader',
+        forwardRef: 'uploader',
     });
 });